requirements.txt
numpy==1.26.4
pygame==2.6.1
PyYAML==6.0.1
